editoast: move SwitchCache to infra_cache
