+++ editoast/src/infra_cache/mod.rs
@@ -34,6 +34,7 @@ use crate::infra_cache::object_cache::BufferStopCache;
 use crate::infra_cache::object_cache::DetectorCache;
 use crate::infra_cache::object_cache::OperationalPointCache;
 use crate::infra_cache::object_cache::SignalCache;
+use crate::infra_cache::object_cache::SwitchCache;
 use crate::infra_cache::operation::CacheOperation;
 use crate::infra_cache::operation::RailjsonObject;
 use crate::map::BoundingBox;
