+++ editoast/src/infra_cache/object_cache.rs
@@ -3,12 +3,14 @@ mod detector_cache;
 mod electrification_cache;
 mod operational_point_cache;
 mod signal_cache;
+mod switch_cache;
 mod switch_type_cache;
 
 pub use buffer_stop_cache::BufferStopCache;
 pub use detector_cache::DetectorCache;
 pub use operational_point_cache::OperationalPointCache;
 pub use signal_cache::SignalCache;
+pub use switch_cache::SwitchCache;
 
 cfg_if! {
     if #[cfg(test)] {
